requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
python-jose[cryptography]
python-multipart
bcrypt==4.0.1
passlib==1.7.4